Install packages automatically and prevent ugly hashes.
The idea is to get a fully configured emacs when loading for the first
time with minimal user involvement. The confirmation to install
selected packages remains just as a sanity check. The call to
load-theme specifies the :no-confirm argument since it's reasonable to
expect that if the package is selected, it's ok to run it's lisp code.